requirements.txt
flask
pymongo
python-dotenv
pytesseract
sympy
google-generativeai